refactor: changing Member class from service to data
