+++ requirements.txt
@@ -4,3 +4,4 @@ isort==5.12.0
 reportlab==4.0.0
 discord.py==2.2.3
 validate_email_address==1
+py3dns==3.2.1
