+++ src/data/__init__.py
@@ -12,6 +12,6 @@ Package Structure
 
 
 from .coordinator_data import CoordinatorData
-from .member_data import MemberData
+from .member_data import Member, MemberData
 from .project_data import ProjectData
 from .student_data import StudentData
